ai-backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
redis
pdfminer.six
nltk
pydantic
